Need to write the table header for the fit results as well - done but requires a bit of tidying up.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@4995 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/current/parametrisation/OptimisedGrower.cc
@@ -75,7 +75,7 @@ namespace askap {
       // this->yObj = this->itsObj->getYPeak();
       this->xObj = int(object->getXCentroid());
       this->yObj = int(object->getYCentroid());
-      float zObj = int(object->getZCentroid());
+      int   zObj = int(object->getZCentroid());
       //   std::cout << "Central position = ("<<this->xObj << ","<<this->yObj <<")\n";
 
       this->findEllipse();
